refactor: adapt to changes from main merge, remove rocket
