+++ common/Cargo.toml
@@ -8,12 +8,11 @@ description = "Common library for the shuttle platform (https://www.shuttle.rs/)
 
 [dependencies]
 anyhow = { version = "1.0.62", optional = true }
-chrono = { version = "0.4.22", features = ["serde"], optional = true }
+chrono = { version = "0.4.22", features = ["serde"] }
 comfy-table = { version = "6.1.0", optional = true }
 crossterm =  { version = "0.25.0", optional = true }
 once_cell = "1.13.1"
 regex = { version = "1.6.0", optional = true }
-rocket = "0.5.0-rc.2" # todo: remove this dep and debug the error from shuttle_service::log in codegen
 rustrict = "0.5.0"
 semver = { version = "1.0.13", optional = true }
 serde = "1.0.137"
@@ -25,5 +24,5 @@ uuid = { version = "1.1.1", features = ["v4", "serde"] }
 [features]
 default = ["models", "version"]
 
-models = ["comfy-table", "chrono", "crossterm", "serde_json"]
+models = ["comfy-table", "crossterm", "serde_json"]
 version = ["anyhow", "semver", "regex"]
