+++ common/src/deployment.rs
@@ -1,0 +1,16 @@
+use serde::{Deserialize, Serialize};
+use strum::Display;
+
+#[derive(Clone, Debug, Deserialize, Display, Serialize)]
+#[serde(rename_all = "lowercase")]
+#[strum(serialize_all = "lowercase")]
+pub enum State {
+    Queued,
+    Building,
+    Built,
+    Running,
+    Completed,
+    Stopped,
+    Crashed,
+    Unknown,
+}
