+++ common/src/log.rs
@@ -1,0 +1,49 @@
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+
+use crate::deployment::State;
+
+pub const STATE_MESSAGE: &str = "NEW STATE";
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct Item {
+    pub id: Uuid,
+    pub timestamp: DateTime<Utc>,
+    pub state: State,
+    pub level: Level,
+    pub file: Option<String>,
+    pub line: Option<u32>,
+    pub target: String,
+    pub fields: Vec<u8>,
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct BuildLogStream {
+    pub id: Uuid,
+    pub state: State,
+    pub message: Option<String>,
+    pub timestamp: DateTime<Utc>,
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(rename_all = "lowercase")]
+pub enum Level {
+    Trace,
+    Debug,
+    Info,
+    Warn,
+    Error,
+}
+
+impl From<&tracing::Level> for Level {
+    fn from(level: &tracing::Level) -> Self {
+        match *level {
+            tracing::Level::ERROR => Self::Error,
+            tracing::Level::WARN => Self::Warn,
+            tracing::Level::INFO => Self::Info,
+            tracing::Level::DEBUG => Self::Debug,
+            tracing::Level::TRACE => Self::Trace,
+        }
+    }
+}
