+++ common/src/models/log.rs
@@ -1,44 +1,11 @@
 use std::fmt::Display;
 use std::fmt::Write;
 
-use chrono::{DateTime, Local, Utc};
+use chrono::{DateTime, Local};
 use crossterm::style::{StyledContent, Stylize};
-use serde::{Deserialize, Serialize};
-use uuid::Uuid;
 
-use crate::models::deployment::State;
-
-pub const STATE_MESSAGE: &str = "NEW STATE";
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct BuildLogStream {
-    pub id: Uuid,
-    pub state: State,
-    pub message: Option<String>,
-    pub timestamp: DateTime<Utc>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct Item {
-    pub id: Uuid,
-    pub timestamp: DateTime<Utc>,
-    pub state: State,
-    pub level: Level,
-    pub file: Option<String>,
-    pub line: Option<u32>,
-    pub target: String,
-    pub fields: Vec<u8>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Eq, PartialEq)]
-#[serde(rename_all = "lowercase")]
-pub enum Level {
-    Trace,
-    Debug,
-    Info,
-    Warn,
-    Error,
-}
+use crate::log::{Item, Level};
+use crate::STATE_MESSAGE;
 
 impl Display for Item {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
@@ -101,18 +68,6 @@ impl Level {
             Level::Info => " INFO".green(),
             Level::Warn => " WARN".yellow(),
             Level::Error => "ERROR".red(),
-        }
-    }
-}
-
-impl From<&tracing::Level> for Level {
-    fn from(level: &tracing::Level) -> Self {
-        match *level {
-            tracing::Level::ERROR => Self::Error,
-            tracing::Level::WARN => Self::Warn,
-            tracing::Level::INFO => Self::Info,
-            tracing::Level::DEBUG => Self::Debug,
-            tracing::Level::TRACE => Self::Trace,
         }
     }
 }
