+++ deployer/src/deployment/deploy_layer.rs
@@ -21,7 +21,7 @@
 
 use chrono::{DateTime, Utc};
 use serde_json::{json, Value};
-use shuttle_common::models::{deployment, log::BuildLogStream};
+use shuttle_common::log::BuildLogStream;
 use shuttle_common::STATE_MESSAGE;
 use std::{net::SocketAddr, str::FromStr};
 use tracing::{error, field::Visit, span, warn, Metadata, Subscriber};
@@ -71,19 +71,19 @@ impl Log {
     pub fn to_stream_log(&self) -> Option<BuildLogStream> {
         let (state, message) = match self.r#type {
             LogType::State => match self.state {
-                State::Queued => Some((deployment::State::Queued, None)),
-                State::Building => Some((deployment::State::Building, None)),
-                State::Built => Some((deployment::State::Built, None)),
-                State::Running => Some((deployment::State::Running, None)),
-                State::Completed => Some((deployment::State::Completed, None)),
-                State::Stopped => Some((deployment::State::Stopped, None)),
-                State::Crashed => Some((deployment::State::Crashed, None)),
-                State::Unknown => Some((deployment::State::Unknown, None)),
+                State::Queued => Some((shuttle_common::deployment::State::Queued, None)),
+                State::Building => Some((shuttle_common::deployment::State::Building, None)),
+                State::Built => Some((shuttle_common::deployment::State::Built, None)),
+                State::Running => Some((shuttle_common::deployment::State::Running, None)),
+                State::Completed => Some((shuttle_common::deployment::State::Completed, None)),
+                State::Stopped => Some((shuttle_common::deployment::State::Stopped, None)),
+                State::Crashed => Some((shuttle_common::deployment::State::Crashed, None)),
+                State::Unknown => Some((shuttle_common::deployment::State::Unknown, None)),
             },
             LogType::Event => match self.state {
                 State::Building => {
                     let msg = extract_message(&self.fields)?;
-                    Some((deployment::State::Building, Some(msg)))
+                    Some((shuttle_common::deployment::State::Building, Some(msg)))
                 }
                 _ => None,
             },
