+++ deployer/src/persistence/log.rs
@@ -1,6 +1,6 @@
 use chrono::{DateTime, Utc};
 use serde_json::Value;
-use shuttle_common::models::{deployment, log::BuildLogStream};
+use shuttle_common::log::BuildLogStream;
 use shuttle_common::STATE_MESSAGE;
 use uuid::Uuid;
 
@@ -33,14 +33,14 @@ impl Log {
         let (state, message) = if let Value::String(str_value) = &self.fields {
             if str_value == STATE_MESSAGE {
                 match self.state {
-                    State::Queued => Some((deployment::State::Queued, None)),
-                    State::Building => Some((deployment::State::Building, None)),
-                    State::Built => Some((deployment::State::Built, None)),
-                    State::Running => Some((deployment::State::Running, None)),
-                    State::Completed => Some((deployment::State::Completed, None)),
-                    State::Stopped => Some((deployment::State::Stopped, None)),
-                    State::Crashed => Some((deployment::State::Crashed, None)),
-                    State::Unknown => Some((deployment::State::Unknown, None)),
+                    State::Queued => Some((shuttle_common::deployment::State::Queued, None)),
+                    State::Building => Some((shuttle_common::deployment::State::Building, None)),
+                    State::Built => Some((shuttle_common::deployment::State::Built, None)),
+                    State::Running => Some((shuttle_common::deployment::State::Running, None)),
+                    State::Completed => Some((shuttle_common::deployment::State::Completed, None)),
+                    State::Stopped => Some((shuttle_common::deployment::State::Stopped, None)),
+                    State::Crashed => Some((shuttle_common::deployment::State::Crashed, None)),
+                    State::Unknown => Some((shuttle_common::deployment::State::Unknown, None)),
                 }
             } else {
                 None
@@ -49,7 +49,7 @@ impl Log {
             match self.state {
                 State::Building => {
                     let msg = extract_message(&self.fields)?;
-                    Some((deployment::State::Building, Some(msg)))
+                    Some((shuttle_common::deployment::State::Building, Some(msg)))
                 }
                 _ => None,
             }
@@ -79,7 +79,7 @@ impl From<Log> for shuttle_common::LogItem {
     }
 }
 
-impl From<Level> for shuttle_common::models::log::Level {
+impl From<Level> for shuttle_common::log::Level {
     fn from(level: Level) -> Self {
         match level {
             Level::Trace => Self::Trace,
